chore(core): yet more scrutinizer issues

+++ mod/blog/start.php
@@ -155,8 +155,8 @@ function blog_archive_menu_setup($hook, $type, $return, $params) {
 	
 	$years = [];
 	foreach ($dates as $date) {
-		$timestamplow = mktime(0, 0, 0, substr($date, 4, 2), 1, substr($date, 0, 4));
-		$timestamphigh = mktime(0, 0, 0, ((int) substr($date, 4, 2)) + 1, 1, substr($date, 0, 4));
+		$timestamplow = mktime(0, 0, 0, (int) substr($date, 4, 2), 1, (int) substr($date, 0, 4));
+		$timestamphigh = mktime(0, 0, 0, ((int) substr($date, 4, 2)) + 1, 1, (int) substr($date, 0, 4));
 
 		$year = substr($date, 0, 4);
 		if (!in_array($year, $years)) {
